added employee DB with all CRUD functionalities

+++ src/app/components/candidate-list/candidate-list.component.ts
@@ -20,6 +20,9 @@ export class CandidateListComponent implements OnInit {
   }
 
   deleteCandidate(id: number) {
-    this.candidateService.deleteCandidate(id).subscribe(() => this.loadCandidates());
+    const confirm = window.confirm("Are you sure?");
+    if(confirm){
+      this.candidateService.deleteCandidate(id).subscribe(() => this.loadCandidates());
+    }
   }
 }
